midterm/midterm_requirements.txt
numpy
pandas
matplotlib
scikit-learn
scipy
torch
torchvision
tqdm
pillow
h5py
